# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Reserva de Espacios",
        "description": "API para gestionar reservas de espacios",
        "version": "1.0.0"
    },
    "paths": {
        "/api/reservations": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Obtener todas las reservaciones",
                "operationId": "fbc6050d9f7fdb07075d3aaeb157ffa9",
                "responses": {
                    "200": {
                        "description": "Lista de reservaciones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Reservation"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Crear una nueva reservación",
                "operationId": "01e3d21dc330a31de025203bc142ffa3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreReservationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reservación creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Reserva creada con éxito"
                                        },
                                        "reservation": {
                                            "$ref": "#/components/schemas/Reservation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al crear la reservación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error al crear la reserva"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/reservations/{id}": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Obtener una reservación específica",
                "operationId": "a226b43b2a6a846245e658d01bf00720",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la reservación",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la reservación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reservation"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Reservación no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Reservation not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Actualizar una reservación existente",
                "operationId": "02fa683a9f80a0563596fb06685bcbc7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la reservación",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreReservationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reservación actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Reserva actualizada con éxito"
                                        },
                                        "reservation": {
                                            "$ref": "#/components/schemas/Reservation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Eliminar una reservación",
                "operationId": "9b6641e4ded1ff7adfce352d589b85e9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la reservación",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reservación eliminada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Reservation deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Reservación no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Reservation not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/spaces": {
            "get": {
                "tags": [
                    "Espacios"
                ],
                "summary": "Obtener lista de espacios",
                "operationId": "e17708e9e7050fff0a684603393f619b",
                "responses": {
                    "200": {
                        "description": "Lista de espacios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Space"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Espacios"
                ],
                "summary": "Crear un nuevo espacio",
                "operationId": "8bbbe73fe6d28fc7aadf84f49791569b",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SpaceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Espacio creado exitosamente"
                    }
                }
            }
        },
        "/api/spaces/{id}": {
            "get": {
                "tags": [
                    "Espacios"
                ],
                "summary": "Obtener un espacio específico",
                "operationId": "42a320419e7c2f110e1cf34132f9e26a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del espacio",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del espacio",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Space"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Espacio no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Espacios"
                ],
                "summary": "Actualizar un espacio existente",
                "operationId": "ad0300609c73b94572ef9b932a2fc6db",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del espacio",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SpaceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Espacio actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Space"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Espacios"
                ],
                "summary": "Eliminar un espacio",
                "operationId": "5a5eda39c31c16b6218dc227d58898a1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del espacio",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Espacio eliminado con éxito"
                    },
                    "404": {
                        "description": "Espacio no encontrado"
                    },
                    "409": {
                        "description": "No se puede eliminar el espacio porque tiene reservas asociadas"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Reservation": {
                "required": [
                    "user_id",
                    "space_id",
                    "start_time",
                    "end_time"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "space_id": {
                        "type": "integer"
                    },
                    "start_time": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "end_time": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Space": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "capacity": {
                        "type": "integer"
                    },
                    "location": {
                        "type": "string"
                    },
                    "is_available": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "ReservationSchema": {
                "required": [
                    "id",
                    "user",
                    "space",
                    "start_time",
                    "end_time"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "space_id": {
                        "type": "integer"
                    },
                    "event_name": {
                        "type": "string"
                    },
                    "start": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "end": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "space": {
                        "$ref": "#/components/schemas/Space"
                    }
                },
                "type": "object"
            },
            "SpaceRequest": {
                "title": "Space Request",
                "description": "Esquema para la solicitud de creación de un espacio",
                "required": [
                    "name",
                    "capacity"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Sala de conferencias"
                    },
                    "capacity": {
                        "type": "integer",
                        "example": 50
                    },
                    "description": {
                        "type": "string",
                        "example": "Sala equipada para conferencias y presentaciones"
                    },
                    "location": {
                        "type": "string",
                        "example": "Edificio principal, piso 2"
                    },
                    "amenities": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "proyector",
                            "pizarra",
                            "wifi"
                        ]
                    }
                },
                "type": "object"
            },
            "StoreReservationRequest": {
                "title": "Store Reservation Request",
                "required": [
                    "space_id",
                    "event_name",
                    "start",
                    "end"
                ],
                "properties": {
                    "space_id": {
                        "type": "integer"
                    },
                    "event_name": {
                        "type": "string"
                    },
                    "start": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "end": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Reservations",
            "description": "Operaciones relacionadas con reservaciones"
        },
        {
            "name": "Espacios",
            "description": "Operaciones relacionadas con espacios"
        }
    ]
}